Network.HTTP.Server.Logger

Description

Documentation

data Logger

A type used by the server to report various events. Useful for debugging.

Constructors

Logger

Fields

logInfo :: Int -> String -> IO ()

logDebug :: String -> IO ()

logError :: String -> IO ()

logWarning :: String -> IO ()

getLog :: Maybe Int -> (LogType -> Bool) -> IO [LogItem]

stdLogger :: Logger

A logger that uses the standard output and standard error. Text is UTF8 encoded.

quietLogger :: Logger

A logger that does not report anything.

utf8Logger :: Handle -> Handle -> Logger

A logger that uses the given handles for output and errors.

data LogItem

Constructors

LogItem

Fields

item_type :: LogType

item_data :: String

data LogType

Constructors

Error

Warning

Debug

Info Int

showLogItem :: LogItem -> String

readLogItem :: String -> Maybe LogItem

filterLog :: Maybe Int -> (LogType -> Bool) -> [LogItem] -> [LogItem]